Python_API_doc/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
app==0.0.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.4
azure-identity==1.14.0
azure-search-documents==11.4.0b8
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
cryptography==41.0.4
dataclasses-json==0.6.0
debugpy==1.8.0
decorator==5.1.1
executing==1.2.0
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.3.1
jupyter_core==5.3.1
langchain==0.0.300
langsmith==0.0.40
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
msal==1.24.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numexpr==2.8.6
numpy==1.26.0
oauthlib==3.2.2
openai==0.28.0
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
portalocker==2.8.2
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
PyJWT==2.8.0
PyPDF2==3.0.1
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
stack-data==0.6.2
tenacity==8.2.3
tiktoken==0.5.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.5
waitress==2.1.2
wcwidth==0.2.6
Werkzeug==2.3.7
yarl==1.9.2
